Refactor chat command models and update architecture documentation
- Moved `ChatInput` and `ChatWsCommand` models from `Application.Abstractions` to `Infrastructure/CapabilityApi`, clarifying the boundary between host protocols and application contracts.
- Introduced `ChatCapabilityMessageTypes` for better organization of message types related to chat commands.
- Updated documentation to reflect changes in model locations and architecture layers, ensuring consistency and clarity in the Workflow LLM streaming architecture.

These changes aim to enhance the structure and maintainability of the chat command handling within the architecture.
