--- src/workflow/Aevatar.Workflow.Application.Abstractions/Runs/WorkflowCapabilityMessageTypes.cs
@@ -1,6 +1,0 @@
-namespace Aevatar.Workflow.Application.Abstractions.Runs;
-
-public static class WorkflowCapabilityMessageTypes
-{
-    public const string ChatCommand = "chat.command";
-}
